Add DateTimeRangeFilter and DateRangeFilter
AbstractDateFilter has been refactored to support range queries
similar to the implementation of sonata-project/doctrine-orm-admin-bundle.

The option input_type has been removed since it was not used.

Prior to this change when filtering using a `DateFilter` or
`DateTimeFilter` with `equals` type, it created a range
query internally. Now you have to use `DateRangeFilter` or
`DateTimeRangeFilter` and provide a range of dates.